layers-store - Fix Bil2diva min value.
git-svn-id: http://alageospatialportal.googlecode.com/svn/trunk/layers-store@2185 73def43e-0e3b-11df-a5af-b38b39d06e06

+++ src/main/java/org/ala/layers/util/Bil2diva.java
@@ -265,7 +265,7 @@ public class Bil2diva {
     }
 
     static void updateMinMax(double[] minmax, double d, double missingValue) {
-        if (d != missingValue) {
+        if ((float) d != (float) missingValue) {
             if (Double.isNaN(minmax[0])) {
                 minmax[0] = d;
                 minmax[1] = d;
